+++ .idea/modules.xml
@@ -2,9 +2,9 @@
 <project version="4">
   <component name="ProjectModuleManager">
     <modules>
-      <module fileurl="file://$PROJECT_DIR$/flutter_datawedge.iml" filepath="$PROJECT_DIR$/flutter_datawedge.iml" />
       <module fileurl="file://$PROJECT_DIR$/android/flutter_datawedge_android.iml" filepath="$PROJECT_DIR$/android/flutter_datawedge_android.iml" />
       <module fileurl="file://$PROJECT_DIR$/example/android/flutter_datawedge_example_android.iml" filepath="$PROJECT_DIR$/example/android/flutter_datawedge_example_android.iml" />
+      <module fileurl="file://$PROJECT_DIR$/flutter_datawedge_plus.iml" filepath="$PROJECT_DIR$/flutter_datawedge_plus.iml" />
     </modules>
   </component>
 </project>
