+++ README.md
@@ -1,11 +1,12 @@
-# flutter_datawedge
 
-[![pub package](https://img.shields.io/pub/v/flutter_datawedge.svg)](https://pub.dev/packages/flutter_datawedge)
+# flutter_datawedge_plus
+
+[![pub package](https://img.shields.io/pub/v/flutter_datawedge_plus.svg)](https://pub.dev/packages/flutter_datawedge_plus)
 
 A Flutter plugin communicate with DataWedge scanners
 
 <p align="center">
- <img width="220px" src="https://raw.githubusercontent.com/rafaeljustinox/flutter_datawedge/main/.github/images/app.png" align="center" alt="Example" />
+ <img width="220px" src="https://raw.githubusercontent.com/circus-kitchens/flutter_datawedge_plus/main/.github/images/app.png" align="center" alt="Example" />
 </p>
 
 ## Getting Started (under development)
