+++ android/build.gradle
@@ -1,4 +1,4 @@
-group 'com.justino.flutter_datawedge'
+group 'com.circuskitchens.flutter_datawedge_plus'
 version '1.0-SNAPSHOT'
 
 buildscript {
