+++ android/settings.gradle
@@ -1,1 +1,1 @@
-rootProject.name = 'flutter_datawedge'
+rootProject.name = 'flutter_datawedge_plus'
