+++ android/src/main/AndroidManifest.xml
@@ -1,3 +1,3 @@
 <manifest xmlns:android="http://schemas.android.com/apk/res/android"
-  package="com.justino.flutter_datawedge">
+  package="com.circuskitchens.flutter_datawedge_plus">
 </manifest>
