+++ android/src/main/kotlin/com/circuskitchens/flutter_datawedge_plus/DWInterface.kt
@@ -1,4 +1,4 @@
-package com.justino.flutter_datawedge
+package com.circuskitchens.flutter_datawedge_plus
 
 import android.content.Context
 import android.content.Intent
