+++ android/src/main/kotlin/com/circuskitchens/flutter_datawedge_plus/Scan.kt
@@ -1,4 +1,4 @@
-package com.justino.flutter_datawedge
+package com.circuskitchens.flutter_datawedge_plus
 
 import androidx.annotation.NonNull
 import org.json.JSONObject
