+++ android/src/main/kotlin/com/circuskitchens/flutter_datawedge_plus/consts/MyIntents.kt
@@ -1,0 +1,7 @@
+package com.circuskitchens.flutter_datawedge_plus.consts
+
+class MyIntents {
+  companion object {
+    const val SCAN_EVENT_INTENT_ACTION ="com.circuskitchens.flutter_datawedge_plus.SCAN_EVENT"
+  }
+}
