--- android/src/main/kotlin/com/justino/flutter_datawedge/consts/MyIntents.kt
@@ -1,7 +1,0 @@
-package com.justino.flutter_datawedge.consts
-
-class MyIntents {
-  companion object {
-    const val SCAN_EVENT_INTENT_ACTION ="com.justino.flutter_datawedge.SCAN_EVENT"
-  }
-}
