+++ example/android/app/build.gradle
@@ -34,7 +34,7 @@ android {
 
     defaultConfig {
         // TODO: Specify your own unique Application ID (https://developer.android.com/studio/build/application-id.html).
-        applicationId "com.justino.flutter_datawedge_example"
+        applicationId "com.circuskitchens.flutter_datawedge_plus_example"
         minSdkVersion 16
         targetSdkVersion 30
         versionCode flutterVersionCode.toInteger()
