+++ example/android/app/src/main/kotlin/com/justino/flutter_datawedge_example/MainActivity.kt
@@ -1,4 +1,4 @@
-package com.justino.flutter_datawedge_example
+package com.circuskitchens.flutter_datawedge_plus_example
 
 import io.flutter.embedding.android.FlutterActivity
