+++ example/test/widget_test.dart
@@ -6,9 +6,8 @@
 // tree, read text, and verify that the values of widget properties are correct.
 
 import 'package:flutter/material.dart';
+import 'package:flutter_datawedge_plus_example/main.dart';
 import 'package:flutter_test/flutter_test.dart';
-
-import 'package:flutter_datawedge_example/main.dart';
 
 void main() {
   testWidgets('Verify Platform version', (WidgetTester tester) async {
@@ -18,8 +17,7 @@ void main() {
     // Verify that platform version is retrieved.
     expect(
       find.byWidgetPredicate(
-        (Widget widget) => widget is Text &&
-                           widget.data!.startsWith('Running on:'),
+        (Widget widget) => widget is Text && widget.data!.startsWith('Running on:'),
       ),
       findsOneWidget,
     );
