+++ pubspec.yaml
@@ -1,8 +1,9 @@
-name: flutter_datawedge
+name: flutter_datawedge_plus
 description: A Flutter plugin communicate with DataWedge scanners.
-version: 0.0.4
-homepage: https://github.com/rafaeljustinox/flutter_datawedge
-repository: https://github.com/rafaeljustinox/flutter_datawedge
+version: 1.0.0
+homepage: https://github.com/circuskitchens/flutter_datawedge_plus
+repository: https://github.com/circuskitchens/flutter_datawedge_plus
+publish_to: 'none' # Remove this line if you wish to publish to pub.dev
 
 environment:
   sdk: ">=2.18.0 <3.0.0"
@@ -18,12 +19,8 @@ dev_dependencies:
 
 
 flutter:
-  # This section identifies this Flutter project as a plugin project.
-  # The 'pluginClass' and Android 'package' identifiers should not ordinarily
-  # be modified. They are used by the tooling to maintain consistency when
-  # adding or updating assets for this project.
   plugin:
     platforms:
       android:
-        package: com.justino.flutter_datawedge
+        package: com.circuskitchens.flutter_datawedge_plus
         pluginClass: FlutterDatawedgePlugin
